ToSMR: changed spanLength remainingLength back to Double, cuz CLLocation uses Double for initialization

+++ ToSMR/Arrival.swift
@@ -23,10 +23,10 @@ public struct Arrival {
     public let nextStopId: Int
     
     /// remainingLength — оставшаяся часть пути (в метрах);
-    public let remainingLength: Float
+    public let remainingLength: Double
     
     /// remainingLength — оставшаяся часть пути (в метрах);
-    public let spanLength: Float
+    public let spanLength: Double
     
     /// requestedStopId — классификаторный номер остановки, для которой запрошен прогноз (помогает разобраться при запросе на несколько остановок сразу);
     public let requestedStopId: Int
